Remove rounding and uneeded type conversion
For most cases, converting from floating-point to integer just requires
the value to be truncated. The conversion function was called for every
operation, which is uneeded. If a explicit method for conversion is
needed (ceiling or rounding) this should be done using one of the
functions to convert floating-point vector to integer vector
`vec*_ceil`, `vec*_round` or `vec*_floor`. The type conversion was also
uneeded in most cases, so it was removed.